\name{SubscaleExplorer-package}
\alias{SubscaleExplorer-package}
\alias{SubscaleExplorer}
\docType{package}
\title{SubscaleExplorer Information}
\packageTitle{SubscaleExplorer}
\description{
\packageDescription{SubscaleExplorer}
}\details{}

\packageDESCRIPTION{SubscaleExplorer}
\packageIndices{SubscaleExplorer}
\author{
\packageAuthor{SubscaleExplorer}

\packageMaintainer{SubscaleExplorer}
}
\references{}

\keyword{ package }
\seealso{}
\examples{
}