\name{SubscaleExplorer-package}
\alias{SubscaleExplorer-package}
\alias{SubscaleExplorer}
\docType{package}
\title{SubscaleExplorer Information}
\packageTitle{SubscaleExplorer}
\description{
\packageDescription{SubscaleExplorer}
}\details{}

\packageDESCRIPTION{SubscaleExplorer}
\packageIndices{SubscaleExplorer}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~

\author{
\packageAuthor{SubscaleExplorer}

Maintainer: \packageMaintainer{}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
