\name{SubscaleExplorer-package}
\alias{SubscaleExplorer-package}
\alias{SubscaleExplorer}
\docType{package}
\title{
\packageTitle{SubscaleExplorer}
}
\description{
\packageDescription{SubscaleExplorer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SubscaleExplorer}
\packageIndices{SubscaleExplorer}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SubscaleExplorer}

Maintainer: \packageMaintainer{SubscaleExplorer}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
