\name{SubscaleExplorer-package}
\alias{SubscaleExplorer-package}
\alias{SubscaleExplorer}
\docType{package}
\title{SubscaleExplorer}
\description{SubscaleExplorer}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{SubscaleExplorer}
\packageIndices{SubscaleExplorer}}
\author{}

\references{}

\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
